CSV to Binary for Efficient Storage
The goal of this project is to implement a program that parses an ASCII
text-based CSV data file and then transforms that same information to a
memory-efficient format and writes it to a binary data file.
The new binary data file also has multiple index files for efficient
record retrieval of values from fields other than the primary key.
This program operates entirely from the command line (no GUI).